Added 16-bit heightmap support. Added support for image formats PNG, TGA, and TIFF. Added marker highlighting.
git-svn-id: http://arkanar/svn/Perimeter%20Map%20Compiler@24 e62d0a23-8ff8-fb48-9cb0-22a7500eb68d

+++ trunk/compiler/stat wnd.h
@@ -42,7 +42,7 @@ class StatWnd : public PanelWindow
 // interface
 public:
 	bool Create(HWND parent_wnd, const RECT &window_rect);
-	void SetAverageHeight(uint height);
+	void SetAverageHeight(float height);
 	void SetAverageColour(COLORREF colour);
 // message handlers
 private:
